requirements.txt
anthropic
fastapi
uvicorn
chromadb
beautifulsoup4
requests
streamlit
python-dotenv
pydantic
